Refactor logging and model loading in server.js (to focuson embed)
Refactor logging and improve model loading logic in server.js.

this edit fixes errors in the last and focuses on embed models. still working on differentiation between embedding and non-embedding models, so for now this version will focus on embedding models and ignore others. which allows you to use this proxy to distribute embedding specifically while still allowing others to be present / loaded.